+++ src/main/java/Task.java
@@ -1,0 +1,5 @@
+import java.util.ArrayList;
+
+public class Tasks {
+
+}
